Only strip leading digits when using `to_ruby_method`. (#66)
When about to release 2.0.0, while testing its integration with an
application, I noticed a regression in `word_chars!` where it would
strip all digits from a string when using `to_slug`. This also didn't
match the docs for the method, as described in #65.

After allowing it to preserve digits this broke `to_ruby_method` because
it was able to create strings like `1` which would result in an invalid
method name. I introduced a new method `strip_leading_digits!` which
takes care of this requirement.

Fixes #65.