Make the build reproducible
Whilst working on the Reproducible Builds effort [0], I noticed that
babosa could not be built reproducibly.

This is because the __dir__ gets expanded when building/rendering the .gemspec
file which is then shipped in the binary package. As this path is dependent on
the path in which you build babosa, this makes the build unreproducible.

I think that cert_chain is only needed if signing_key is needed, so moving it
under the same conditional seems to be safe. The goal, however, is to prevent
cert_chain from being included in regular, "normal", builds of the package.

I originally filed this in Debian as bug #1041840 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1041840